Create new components for Actions Modal and stuff
